rename: 공통 컴포넌트 위계 구조에 의한 분리
